requirements.txt
requests>=2.25.1
Pillow>=8.2.0
pytesseract>=0.3.8
pandas>=1.2.4
numpy>=1.20.3
tqdm>=4.61.0
python-dateutil>=2.8.1
